Fix API breakage in `Command::spawn`
`Command::spawn` should set `stdout` and `stderr` to `Stdio::null()` if
they are not set.

However, due to the fact that `Command::output` set both to
`Stdio::piped()`, the default value is overwritten, thus it has to be
explicitly set.

I thought of setting them back to `Stdio::null()`, however due to the
fact that `Future` can be cancelled, it does not guarantee that they
will be set back.

Signed-off-by: Jiahao XU <Jiahao_XU@outlook.com>